+++ inc/LNS/Parallel/TimeSpaceAStarPlanner.h
@@ -16,7 +16,7 @@ class TimeSpaceAStarPlanner {
 public:
     Instance & instance;
     std::shared_ptr<HeuristicTable> HT;
-    std::shared_ptr<vector<int> > weights;
+    std::shared_ptr<vector<float> > weights;
 
     boost::heap::pairing_heap<TimeSpaceAStarState*, boost::heap::compare<TimeSpaceAStarState::Compare> > open_list;
     boost::unordered_set<TimeSpaceAStarState*, TimeSpaceAStarState::Hash, TimeSpaceAStarState::Equal> all_states;
@@ -32,7 +32,7 @@ public:
     static const int n_dirs=5; // east, south, west, north, stay
     static const int n_orients=4; // east, south, west, north
 
-    TimeSpaceAStarPlanner(Instance & instance, std::shared_ptr<HeuristicTable> HT, std::shared_ptr<vector<int> > weights);
+    TimeSpaceAStarPlanner(Instance & instance, std::shared_ptr<HeuristicTable> HT, std::shared_ptr<vector<float> > weights);
     void findPath(int start_pos, int start_orient, int goal_pos, ConstraintTable & constraint_table, const TimeLimiter & time_limiter);
     void clear();
     void buildPath(TimeSpaceAStarState * curr, int goal_pos);
