+++ inc/LaCAM2/LaCAM2Solver.hpp
@@ -33,7 +33,7 @@ public:
 
     std::shared_ptr<Graph> G; // graph
     std::shared_ptr<HeuristicTable> HT; // instance
-    std::shared_ptr<std::vector<int> > map_weights; // map weights
+    std::shared_ptr<std::vector<float> > map_weights; // map weights
     // Config next_config;
 
     std::vector<std::vector<int>> action_costs;
@@ -48,12 +48,12 @@ public:
 
     Instance build_instance(const SharedEnvironment & env, std::vector<Path> * precomputed_paths=nullptr);
 
-    int get_action_cost(int pst, int ost, int ped, int oed);
-    int eval_solution(const Instance & instance, const Solution & solution);
+    float get_action_cost(int pst, int ost, int ped, int oed);
+    float eval_solution(const Instance & instance, const Solution & solution);
 
     int get_neighbor_orientation(int loc1,int loc2);
 
-    LaCAM2Solver(const std::shared_ptr<HeuristicTable> & HT, SharedEnvironment * env, std::shared_ptr<std::vector<int> > & map_weights, nlohmann::json & config):
+    LaCAM2Solver(const std::shared_ptr<HeuristicTable> & HT, SharedEnvironment * env, std::shared_ptr<std::vector<float> > & map_weights, nlohmann::json & config):
         HT(HT),
         map_weights(map_weights),
         action_model(env),executor(env),slow_executor(env),
