+++ src/MAPFPlanner.cpp
@@ -150,7 +150,7 @@ void MAPFPlanner::rhcr_config_solver(std::shared_ptr<RHCR::RHCRSolver> & solver,
 std::string MAPFPlanner::load_map_weights(string weights_path) {
     // TODO(rivers): make weights float
     // we have at least 5 weights for a location: right,down,left,up,stay
-    map_weights=std::make_shared<std::vector<int> >(env->rows*env->cols*5,1);
+    map_weights=std::make_shared<std::vector<float> >(env->rows*env->cols*5,1);
     std::string suffix = "all_one";
 
     if (weights_path!=""){
@@ -164,7 +164,7 @@ std::string MAPFPlanner::load_map_weights(string weights_path) {
             }
 
             for (int i=0;i<map_weights->size();++i){
-                (*map_weights)[i]=_weights[i].get<int>();
+                (*map_weights)[i]=_weights[i].get<float>();
             }
             
         }
